server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Edd Salkield <edd@salkield.uk>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
Flask-Admin = "^1.5.6"
Flask-SQLAlchemy = "^2.4.4"
filedepot = "^0.8.0"
pillow = "^7.2.0"
Flask-Login = "^0.5.0"
Flask-WTF = "^0.14.3"
email_validator = "^1.1.1"
werkzeug = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
server = "run:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
